// src/types/RequestPartnerFromAdmin.tsx
export type RequestPartnerFromAdmin = {
    email: string;
    password: string;

    businessRegistrationNumber: string;
    organizationName: string;
    industry: string;
    phoneNumber: string;
    website: string;
    address: string;
    representativeName: string;
    position: string;
    description: string;
    logoUrl: string;
    establishedYear: number;

}